refactor(webapp): move types.ts, env.d.ts, react-table.d.ts into src/types/
